Add error handlings

+++ src/controller/tasks.controller.ts
@@ -15,8 +15,11 @@ export class TaskController {
   get(req: Request, res: Response) {
     this.repo.read().then((data) => {
       const { id } = req.params;
-      const infoId = data.find((item) => item.id === Number(id));
-      res.json(infoId);
+      const getTask = data.find((item) => item.id === Number(id));
+      // eslint-disable-next-line no-unused-expressions
+      getTask === undefined
+        ? res.status(400).send(`<h1>Task id ${id} does not exist</h1>`)
+        : res.json(getTask);
     });
   }
 
@@ -29,16 +32,30 @@ export class TaskController {
   async patch(req: Request, res: Response) {
     const id = Number(req.params.id);
     const currentTask: any = await this.repo.readById(id);
-    const incomingTask = req.body;
-    const updatedTask = Object.assign(currentTask, incomingTask);
-    await this.repo.update(updatedTask);
-    res.send('Task updated');
+    console.log(currentTask);
+
+    if (currentTask === undefined) {
+      res.status(400).send(`<h1>Task id ${id} does not exist</h1>`);
+    }
+
+    if (currentTask !== undefined) {
+      const incomingTask = req.body;
+      const updatedTask = Object.assign(currentTask, incomingTask);
+      await this.repo.update(updatedTask);
+      res.send('Task updated');
+    }
   }
 
   async delete(req: Request, res: Response) {
     const id = Number(req.params.id);
     const deleteTask: any = await this.repo.readById(id);
-    await this.repo.delete(deleteTask);
-    res.send('Task removed');
+    if (deleteTask === undefined) {
+      res.status(400).send(`<h1>Task id ${id} does not exist</h1>`);
+    }
+
+    if (deleteTask !== undefined) {
+      await this.repo.delete(deleteTask);
+      res.send(`Task with id ${id} removed`);
+    }
   }
 }
